requirements-conda.txt
python=3.9
pytorch>=2.0.0
torchvision
torchaudio
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter
ipykernel

